+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### 1.3.2
+- Fixed `AtrributeError` during subscriptions
+
 ### 1.3.1
 - Added ignore cache
 - Re-enabled regex match type